synth_relock_count as log2
This changes synth_relock_count to be log2 + 1 the number of resets. We
mostly care that that relocking is happening none at all, a little bit,
or a lot, and in the a lot case we don't want to overflow the value.
log2 + 1 gives the vague range, and we'd be unlikely to see a value over
~15.

This also resets the value on boot.