Replace RadiosService.recv_queue list with SimpleQueue
In general it's a good idea to use python's queue module when sharing
list-like things across threads, they're built explicitly with threading
in mind. The list here wasn't experiencing racing issues but it's very
easy to accidentally introduce them.

The main problem though, is that the list was being operated as a FIFO,
whereas we wanted a LIFO. This would result in packets arriving in
reversed order when sending multiple packets quickly enough.
Specifically .append() adds an item to the end of the list and .pop()
with no arguments removes an item from the end, not the beginning as
what was intended.

SimpleQueue is explicitly LIFO only, and so can't be misused in this
way.