Restore file upload script random functionality
The previous implementation generated a random bytes sequence and sent
that directly without going through the file system. With a real CFDP
implementation, all the file-related actions need to be done through the
filestore so the random data here should probably be written to disk.
This also makes it easier to check that the source file and resulting
file in the destination are the same.

I had initially been trying to use tempfiles and arranging the arguments
to have everything handled behind the scenes but I came to realize that
it was an overcomplicated solution and what I actually wanted was a
non-temp file opened in exclusive creation mode. This patch is much more
simple than I had initially feared.