Handle unknown remote entity in proxy put
This handles the exception that occurs when a proxy put request requests
transfer to a destination that the proxy does not know about (i.e. it's
not in its MIB). The spec as far as I could see doesn't really consider
this case, and there's no standard error code for it. The least worst
set of error codes I could find ended up being the same as when the file
requested does not exist on the proxy either.