Fixes for EofPdu.unpack(), ack_counter
EofPdu.unpack() wasn't properly unpacking a condition code. This code
was then being passed to AckPdu and getting repacked, but the value was
then over 255, causing it to raise an exception because it couldn't fit
in a byte.

ack_counter was being reset when the FinishedPdu wasn't being responded
to, so the DestHandler would get stuck in a loop sending the PDU
forever, instead of only up to the limit (which is default 2).