Handle sdo writes to type DOMAIN and OCTET_STRING in edl_cmd_shell
In 5f961ed I added DOMAIN handling so that we could send os commands,
but DOMAIN really is supposed to be for file-like things. This enhances
sdo_write to interpret the final argument instead as a filename, which
it then opens and sends.

Additionally os commands really should have been of type OCTET_STRING
but that needs to be fixed in oresat-configs. This adds OCTET_STRING
handling in anticipation of that.