Improve ping script output readability
Previously each ping iteration could print up to three lines, and it was
sometimes easy to miss if the latest iteration succeeded or failed. This
tries to limit each iteration to only a single vertically aligned line,
printing information as it comes in.

While I was here I simplified the implementation:
- By being a little more careful about where and how long blocking
  happens I was able to drop multiple threads. This makes coordinating
  printing lines and shared data much easier
- Removed previous loop timestamp tracking, it didn't seem to work
  anyway?