Simple scrip to print received beacons
Mirroring the EDL scripts this script listens for beacons from the C3
and prints currently just the callsigns and if checksum failed. Other
relevant fields could also be added but all the fields might be too
much.

We've wanted something like this in ground testing on flatclogs/ebv1
where either we didn't want to or couldn't run a full yamcs instance.
This is also handy testing on a local machine for the same reasons.