Fix beacon payload offsets in script
It turns out that when all the test values are zero it's easy to confuse
one zero with another. I had copied the offsets here from the
readthedocs except there were two flaws:
- I copied them verbatim but those also include the header in the
  offsets whereas `ax.payload` does not. All the values need to have 16
  subtracted from them.
- Due to a caching issue (see e6583f in oresat-configs) some values had
  the wrong larger type and so the offsets were off for all later
  entries.

Eventually there should be a beacon.pack()/.unpack() so I don't have to
bother with manually setting offsets, but until then this suffices.