Improve readability of dropped packet text
Revisiting this weeks later with a fresh eye, the text for a dropped
packet is confusing in live use. It looked identical to a received
packet except for an x instead of a ms. The received packet (echo) count
was also confusing unless you knew exactly what was going on.

Removing the echo count from dropped packets removes the point of
confusion and makes the text more visually distinct.