Add beacon pings to edl_ping_loop
Beacon ping is still a ping right? For a handful of reasons (including
regulatory) we need to send beacon pings periodically throughout a pass.
Since this is the current command generator for passes and this does
pings it makes sense to me to add more kinds of pings. This will do it
every 30 seconds/edl pings by default but that could be configurable.

Also it doesn't receive beacons, that's the domain of beacon.py. But it
could...